feat: add env for logs directory
