+++ .env
@@ -2,6 +2,7 @@ NODE_ENV=development
 
 # Set app log level: ALL, INFO, WARN, ERROR
 LOG_LEVEL=all
+LOG_DIR=./logs
 
 # Set app connection
 PORT=4000
